update in rooms

+++ admin/ajax/features_facilities.php
@@ -42,14 +42,21 @@ if(isset($_POST['get_features']))
 }
 
 if(isset($_POST['rem_feature'])){
-    $frm_data=filteration($_POST);
-    $values=[$frm_data['rem_feature']];
+    $frm_data = filteration($_POST);
+    $feature_id = $frm_data['rem_feature'];
+    $values = [$feature_id];
 
-    $q="DELETE FROM `features` WHERE `id`=?";
-    $res=delete($q,$values,'i');
-    echo $res;
+    $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?', $values, 'i');
 
+    if(mysqli_num_rows($check_q) == 0) {
+        $q = "DELETE FROM `features` WHERE `id`=?";
+        $res = delete($q, $values, 'i');
+        echo $res;
+    } else {
+        echo 'room_added';
+    }
 }
+
 
 
 if(isset($_POST['add_facility']))
@@ -106,19 +113,25 @@ if(isset($_POST['rem_facility'])){
     $frm_data=filteration($_POST);
     $values=[$frm_data['rem_facility']];
 
+    $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?', $values, 'i');
 
-    $pre_q="SELECT * FROM `facilities` WHERE `id`=?";
-    $res=select($pre_q,$values,'i');
-    $img =mysqli_fetch_assoc($res);
+    if(mysqli_num_rows($check_q) == 0) {
+        $pre_q="SELECT * FROM `facilities` WHERE `id`=?";
+        $res=select($pre_q,$values,'i');
+        $img =mysqli_fetch_assoc($res);
 
-    if(deleteImage($img['icon'],FACILITIES_FOLDER)){
-        $q="DELETE FROM `facilities` WHERE `id`=?";
-        $res=delete($q,$values,'i');
-        echo $res;
-    }else{
-        echo 0;
+        if(deleteImage($img['icon'],FACILITIES_FOLDER)){
+            $q="DELETE FROM `facilities` WHERE `id`=?";
+            $res=delete($q,$values,'i');
+            echo $res;
+        }else{
+            echo 0;
+        }
     }
-    
+
+    else {
+        echo 'room_added';
+    }
 
 }
